Administrator Functions User Access
These files contain the user interface for the Administrator Functions User Access for the Resnick Alumni Tracking system along with the instructions. The UI and instructions show and explain what the features of the administrator access are and how to add, edit, or remove the access for users of the system.